% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{image.CannyEdges-package}
\alias{image.CannyEdges-package}
\title{Implementation of the Canny Edge Detector for Images}
\description{
Canny Edge Detector for Images. See \url{https://en.wikipedia.org/wiki/Canny_edge_detector}.
Adapted from \url{https://github.com/Neseb/canny}.
}
\seealso{
\link{image_canny_edge_detector}
}
